requirements.txt
pillow==12.1.*
pillow_heif==1.3.*
reportlab==4.4.*